assignment-3-python/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openai==1.30.1
python-dotenv==1.0.1
